requirements.txt
discord.py==2.1.1
Requests==2.31.0
